widen ext_input to 16 bits

+++ src/ant/pin.rs
@@ -161,7 +161,7 @@ impl Pin {
 		PinDefinition {
 			pin: Self::ExtIn,
 			code: "K",
-			size: BYTE,
+			size: DOUBLE,
 			io_type: Some(IoType::Input),
 		},
 		PinDefinition {
