+++ src/world/state.rs
@@ -33,7 +33,7 @@ pub struct WorldState {
 	pub ants: Layers,
 	pub signal_in: u8,
 	pub signal_out: u8,
-	pub ext_input: u8,
+	pub ext_input: u16,
 	pub ext_output: Vec<u16>,
 }
